+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## Unreleased
 
+### Added
+
+- [#4616](https://github.com/pulumi/pulumi-kubernetes/pull/4616) `helm.v4.Chart` now accepts a `kubeVersion` input, equivalent to `helm template --kube-version`, to override the Kubernetes version used for `Capabilities.KubeVersion`.
+
 ## 4.34.2 (September 18, 2026)
 
 ### Fixed